Move JAXB adapters to io.jaxb package
