+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/io/jaxb/JaxbCustomPropertiesAdapter.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.optaplanner.core.impl.util;
+package org.optaplanner.core.impl.io.jaxb;
 
 import java.util.List;
 import java.util.Map;
