+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/io/jaxb/JaxbDurationAdapter.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.optaplanner.core.impl.util;
+package org.optaplanner.core.impl.io.jaxb;
 
 import java.time.Duration;
